Feat: Add React Router
- Removed the boilerplate React homepage
- Created a separate component, for each page we link to. Hello.js and HelloAgain.js
- Renamed the hello.js to Hello.js

I think the issue was that we were trying to route to the same component that the Router was wrapped around. I don't know if my implementation is the most correct solution, but a way around that problem was to give our home page '/' it's own component (Hello.js) to render the new view separate from App.js.