Display items from database in List component
This commit adds `firebase` and `react-firestore` as dependancies so that we can interact with our database.

A List component was created where we'll be able to house all the list's logic and display the list items that get added to the database.

Inside that List component, we've used `FirestoreCollection` which is a component that allows you to interact with a Firestore collection. Using this component, you can access the collection at a given path and provide sort options, perform queries, and paginate data.

This component will setup a listener and update whenever our collection is updated in Firestore.