Test: Add the token generator
We imported the token.js file that was provided in our story. We created a component called NewList to test if we could generate a token and display it to a view.

The NewList component was added to AddItem.js and we were successful in creating and displaying a new token upon reloading the page.