Test: Link to NewList component
When I imported the NewList component in AppRouter.js, I did it incorrectly and had the path './pages/List' instead of './pages/NewList'. Because of that bug, it looked like the routing wasn't working. Now the generated token displays on its own page.

I also added a link to the NewList component on the home page in List.js and a link back home on NewList.js.