Feat: Set up a new shopping list
A shopping list consists of a set of items associated with a user’s token. Tokens can be shared with other users to allow them to co-manage a given list. Creating a new list consists of the following:

  [x] Generate a new, unique token
  [x] Save the token to localStorage
  [x] Show the user the list view

We used the getToken function to generate a new unique token. To associate the token with a list, we save it in local storage. If it is the first time visiting the list a new token is generated. If the user is returning, we get the saved token from local storage. The last criterion was to show the list view. Currently, the list associated with a token is empty. We added placeholder text that shows the current token to demonstrate that the token persists.