Refactor: Change initialToken to a function
I removed the variable used to generate a token and converted the initialToken variable to a function. The behavior didn't change.

`initialToken` is called once when the `token` state variable is initialized. It takes the value stored in local storage if it exists, if not it generates a token with the getToken function.