Fix: Show all items associated with a given token
An index needed to be built so that we could run a compound query. I created the query without using the FirestoreCollection component so that I could get the link from the error message in the console to build the index. I'm not sure if that was necessary or not but that's how I got there.