Move addItem function to listContext.js
I implemented most of the changes Steve suggested. I removed the setTimeout and now the function checks for duplicates as you type.

Does this break UX? I tried entering apple then apple pie. The app will warn that apple is on the list but if you keep typing the warning goes away.