Fix: New list empty token bug
The bug of not being able to add items to a new empty list after switching from an existing list was reintroduced. I think I fixed it by making sure that a new list and new token is generated and saved to local storage if the user enters an empty string in the token field.