Handle the token for global use (#61)
* first commit for PR

* add TokenContext

* Use and set the token with tokenContext.js

I think I've made all the changes needed to have the token stored in just one place in our app. The bug of adding to the wrong list is resolved by using the TokenContext token instead of setting to an empty string.

* Move getting and setting token to its own context

I took the handling the token out of list context and made the changes in the other files so that we always access the token from tokenContext.js or localStorage.

* Fix: Bug in List.js fetching tan existing token

* Update JoinList to use TokenContext

* Remove the database call frrom JoinList.js

* Make sure token is getting fetched and saved to localstorage using the TokeContext functions

* Fix typo so getLocalStorageToken() gets called

* Remove Consumer from export, remove token from getLocalStorageToken parameters

* Convert HomePage to functional component

* Merge token changes

* Fix: New list empty token bug

The bug of not being able to add items to a new empty list after switching from an existing list was reintroduced. I think I fixed it by making sure that a new list and new token is generated and saved to local storage if the user enters an empty string in the token field.

* Bug: The token doesn't clear when no matching list is found

If I make up a token I will get an empty list as expected but at this point I'm also creating a new list with the token I made up. Feature or bug?

* Create displayList function in listContext.js

* Only render the add item form when a valid token is present

I was having problems with adding items to the wrong list and being able to clear the token from a previous list after switching to an invalid token. This solution temporarily breaks being able to start a new list.

* Go to HomePage instead of error message if no valid token

* Clean up unused code

* Start rendering the list on homepage

* space in AddItem so I can pull

* Replace tokenContext with useListToken, remove displayList() from listContext

* Consistently use the useListToken() to get the token

* Add valid token check to addItmem function

* Fix: Allow a new list to be started

If the token is empty when an item is added to a new list a token will be generated on submit

* Move JoinList to Welcome component

* removed unneeded code from HiddenButton.js

* removed unecessary code on list.js

* Fix critical bug on HomePage.js

Co-authored-by: Amber <ajiles91@gmail.com>