Attempt to apply suggested fixes
When one thing moved it started breaking other parts. I tried to apply the suggestions Steve made but the token wasn't saving to local storage in time. When I entered a new token on the join view it didn't save that token and used the autogenerated one. That is a problem when you add an item to the list becuase it starts a new list.

I tried to get around it but adding url parameters but the problem I just mentioned remained. I will stop here and let the team give feedback.