All pages styled
All the pages have a basic level of styling. For usability I also added a delete item function on the shopping list view.

There was a problem where I was losing the token if the addItem screen was the entry to the app. It was because there was no communication with the list in context and we were using that list to check for duplicates.

To fix this I added a call to fetchList when addItem loads. Let me know if this breaks anything.