Make things pretty (#67)
* Add Semantic UI

* Start replacing components on List.js, Add header

* Convert list to Semantic components

* Add theming to App

Theming is set up so that when we decide on fonts, colors, and custom layouts we can update and/or override the default styles using less variables and css.

* Style list

* Add list icon to header, remove header from welcome page

In an earlier commit I added a site header to App.js I removed it so that the welcome page for new lists can have it's own header. I changed some colors also just to play with it. I'm not married to any style choices I made we can still talk about final design choices.

* Add color code to list

* Add delete button to list item

* Change button layout and color

* Change to outline from background color

* Clean up unused functions

* Remove animation from list

* Clean up HomePage.js

* Change Welcome prompt to Semantic

* Simplify welcome prompt

Instead of making the welcome prompt two parts it's not just one view. The join list field and button are responsive.

* Add details dropdown to list items

I added a drop down with some placeholder data and moved the delete button to inside of it. Also, I took the popup off of the delete button action because it was behaving strangely on mobile.

* All pages styled

All the pages have a basic level of styling. For usability I also added a delete item function on the shopping list view.

There was a problem where I was losing the token if the addItem screen was the entry to the app. It was because there was no communication with the list in context and we were using that list to check for duplicates.

To fix this I added a call to fetchList when addItem loads. Let me know if this breaks anything.

* Remove a CSS rule

* Add missing functions to listContext

* Remove unused files, adjust addItem styles

* Center Add Item header

* Remove test content from item details

* Clean up comments and console.log statements, adjust color code logic

* Update .gitignore

* Remove Semantic folder from gitignore

* Try ignoring /src/semantic-ui/site

* Add back theme files

* Clean up unneeded <div>, small changes

I cleaned up a an unneeded div tag, removed some blank lines, and moved the updateItem function closer to purchaseItem in listContext.js.